+++ NEWS.md
@@ -1,9 +1,16 @@
 
 # renv 0.16.0  (UNRELEASED)
 
+* `renv` now builds packages before install, when installing packages from
+  source archives that have not yet been built (as via `R CMD build`). This
+  can make package installation slower, but will ensure packages are installed
+  with vignettes built. (#585)
+  
+
 # renv 0.15.2
 
 * Fixed issues encountered in R CMD check.
+
 
 # renv 0.15.1
 
@@ -27,6 +34,7 @@
 
 * Fixed an issue where `config$cache.symlinks()` would report `NULL` when
   unset. (#906)
+
 
 # renv 0.15.0
 
@@ -167,6 +175,7 @@
   to pass along other optional arguments to the shimmed function correctly.
   (#808)
   
+
 # renv 0.14.0
 
 * `renv` now uses `tools::R_user_dir()` to resolve the default path to the
@@ -281,6 +290,7 @@
   for example, `renv::load()` masking `base::load()`. When set, all usages
   of `renv` APIs must be explicitly qualified with the `renv::` prefix.
   
+
 # renv 0.13.2
 
 * `renv::install("user/repo/subdir with spaces")` now works as expected. (#694)
@@ -303,6 +313,7 @@
 * Fixed an issue where `renv::status()` could provide a misleading message
   for packages which are recorded in the lockfile, but not explicitly
   required by the project. (#684)
+
 
 # renv 0.13.1
 
@@ -427,9 +438,11 @@
 * Fixed an issue where package installation could fail if the `configure.vars`
   option was set to be a named character, rather than a named list. (#609)
 
+
 # renv 0.12.5
 
 * Fixed an issue where `renv` would fail to bootstrap. (#608)
+
 
 # renv 0.12.4
 
@@ -481,6 +494,7 @@
 * Fixed an issue where bootstrapping an older version of `renv` could
   fail if the R repositories had not been appropriately set.
 
+
 # renv 0.12.3
 
 * Fixed an issue where `renv::dependencies()` could give an error if called
@@ -520,9 +534,11 @@
 * When `RENV_PATHS_LIBRARY_ROOT` is set, `renv` will now disambiguate library
   paths based on a hash of the project's path. (#564)
 
+
 # renv 0.12.2
 
 * `renv` no longer errs when running tests with `_R_CHECK_SUGGESTS_ONLY_=false`.
+
 
 # renv 0.12.1
 
@@ -578,6 +594,7 @@
   enabled by setting `options(renv.config.locking.enabled = TRUE)` in an
   appropriate R startup file. (#525)
 
+
 # renv 0.12.0
 
 * `renv` now uses R's internal tar implementation by default on Windows. This is
@@ -614,6 +631,7 @@
 * Fixed an issue where the `RENV_PATHS_PREFIX` environment variable was
   inappropriately normalized when `renv` was loaded. (#465)
   
+
 # renv 0.11.0
 
 * Fixed an issue where `renv::install(..., type = "binary")` would
@@ -670,6 +688,7 @@
 * `renv` now records the `OS_type` reported in a package's `DESCRIPTION` file
   (if any), and ignores packages incompatible with the current operating
   system during restore. (#394)
+
 
 # renv 0.10.0
 
@@ -784,6 +803,7 @@
 * `renv` now tries to place Rtools on the PATH when a package is installed
   with the `install.packages()` hook active. (#335)
 
+
 # renv 0.9.3
 
 * Fixed an issue where attempts to specify `RENV_PATHS_RTOOLS` would
@@ -848,13 +868,16 @@
 * `renv` now updates its local `.gitignore` file, when part of a git repository
   whose git root lives in a parent directory. (#300)
 
+
 # renv 0.9.2
 
 * Fixed an issue in invoking `find` on Solaris.
 
+
 # renv 0.9.1
 
 * Fixed an issue in invoking `cp` on Solaris.
+
 
 # renv 0.9.0
 
@@ -929,6 +952,7 @@
 * `renv::install()` disables staged package install when running with the
   Windows Subsystem for Linux. (#239)
 
+
 # renv 0.8.3
 
 * `renv::dependencies()` gains a new argument `dev`, indicating whether
@@ -973,13 +997,16 @@
 * Fixed an issue where attempts to call `renv::restore()` with the path to the
   lockfile explicitly provided would fail. (#227)
 
+
 # renv 0.8.2
 
 * Further fixes for checks run on CRAN.
 
+
 # renv 0.8.1
 
 * Fixes for checks run on CRAN.
+
 
 # renv 0.8.0